backend/requirements.txt
flask
spyne
psycopg2-binary
PyJWT
python-dotenv
bcrypt
lxml
six